yolo/requirements.txt
# Core Dependencies for Ambulance Detection Dashboard
# Install with: pip install -r requirements.txt

# Deep Learning & Computer Vision
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
opencv-python>=4.5.0
pillow>=8.0.0

# API Framework
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5

# Data Processing
numpy>=1.21.0
pandas>=1.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
pathlib2>=2.3.0
pyyaml>=5.4.0
tqdm>=4.62.0

# Optional: CUDA support (install separately)
# torch with CUDA: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

# Development tools (optional)
# jupyter>=1.0.0
# notebook>=6.4.0

# Production deployment (optional)
# gunicorn>=20.1.0
